+++ src/modules/filter/store/stateFilter.js
@@ -31,6 +31,7 @@
  * @property {String} linkText - link text of the url link to current filter state or empty string if no such link should be rendered
  * @property {Boolean} closeGfi - If it is true, the open gfi window will be closed after new filtering.
  * @property {Boolean} preventAdjust - If it is true, the snippet will not be adjusted.
+ * @property {String} questionLink - The url of the question link.
  */
 
 const state = {
@@ -64,7 +65,8 @@ const state = {
     closeGfi: false,
     preventAdjust: false,
     linkText: "",
-    initialStartupReset: false
+    initialStartupReset: false,
+    questionLink: ""
 };
 
 export default state;